App/MemePhotoApp/requirements.txt
flask
Pillow
requests
git+https://github.com/google/generative-ai-python.git
textblob
pillow
